Restructure ml.inference.
