+++ asapdiscovery-docking/asapdiscovery/docking/scorer_v2.py
@@ -14,7 +14,7 @@ from asapdiscovery.docking.docking_data_validation import (
     DockingResultColsV2 as DockingResultCols,
 )
 from asapdiscovery.docking.docking_v2 import DockingResult
-from asapdiscovery.ml.inference.inference import (
+from asapdiscovery.ml.inference import (
     InferenceBase,
     get_inference_cls_from_model_type,
 )
