+++ asapdiscovery-docking/asapdiscovery/docking/scripts/dock_small_scale_end_to_end.py
@@ -43,7 +43,7 @@ from asapdiscovery.docking import (
     make_docking_result_dataframe,
 )
 from asapdiscovery.docking.docking_data_validation import DockingResultCols
-from asapdiscovery.ml.inference.inference import GATInference, SchnetInference
+from asapdiscovery.ml.inference import GATInference, SchnetInference
 from asapdiscovery.modeling.modeling import protein_prep_workflow
 from asapdiscovery.modeling.schema import (
     MoleculeFilter,
